feat: 레이싱 스코어 객체
